+++ app/Http/Controllers/LA/TicketsController.php
@@ -234,7 +234,7 @@ class TicketsController extends Controller
 				}
 				
 				if(Module::hasAccess("Tickets", "delete")) {
-					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.tickets.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline']);
+					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.tickets.destroy', $data->data[$i][0]], 'method' => 'delete', 'onsubmit'=> 'return confirm("¿Seguro que quieres borrar?")', 'style'=>'display:inline']);
 					$output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
 					$output .= Form::close();
 				}
