+++ app/Http/Controllers/LA/Tickets_repliesController.php
@@ -236,7 +236,7 @@ class Tickets_repliesController extends Controller
 				}
 				
 				if(Module::hasAccess("Tickets_replies", "delete")) {
-					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.tickets_replies.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline']);
+					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.tickets_replies.destroy', $data->data[$i][0]], 'method' => 'delete', 'onsubmit'=> 'return confirm("¿Seguro que quieres borrar?")', 'style'=>'display:inline']);
 					$output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
 					$output .= Form::close();
 				}
