+++ app/Http/Controllers/LA/UsersController.php
@@ -118,7 +118,7 @@ class UsersController extends Controller
 				}
 				
 				if(Module::hasAccess("Users", "delete")) {
-					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.users.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline']);
+					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.users.destroy', $data->data[$i][0]], 'method' => 'delete', 'onsubmit'=> 'return confirm("¿Seguro que quieres borrar?")', 'style'=>'display:inline']);
 					$output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
 					$output .= Form::close();
 				}
